requirements.txt
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.3
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==0.21.0
Werkzeug==2.2.3
zipp==3.15.0 
Flask-SQLAlchemy==2.5.1
pymysql==1.0.2
pytz==2021.1
SQLAlchemy==1.4.22
